Add output panel as an option to show argument hints
Add a preference setting: "tern_argument_hints_type", which can be
set to either "status" or "panel". If "tern_argument_hints" is set
to True, this setting determines how the arghints are shown.
"status" is the default value and uses the old status bar mechanism,
while "panel" uses an output panel to show the arghints.